requirements.txt
streamlit
python-dotenv
huggingface_hub
sqlalchemy
psycopg2-binary